Rearrange files, switch random number generator to xoroshiro128

+++ src/hitables/hitable_list.h
@@ -4,7 +4,7 @@
 #define HITABLE_LIST_H
 
 #include "hitable.h"
-#include "utility/ray.h"
+#include "utility/data_types/ray.h"
 
 class hitable_list
     : public hitable
@@ -30,10 +30,10 @@ inline bool hitable_list::hit(
 {
     hit_record temp_rec;
     bool hit_anything = false;
-    double closest_so_far = t_max;
+    float closest_so_far = t_max;
     for (auto i = 0; i < list_size; i++)
     {
-        if (list[i]->hit(r, t_min, closest_so_far,temp_rec))
+        if (list[i]->hit(r, t_min, closest_so_far, temp_rec))
         {
             hit_anything = true;
             closest_so_far = temp_rec.t;
